Rename file and class names.
