+++ navigation/navigation/src/main/java/Browser.java
@@ -19,18 +19,17 @@
  */
 
 import com.teamdev.jxbrowser.chromium.Browser;
-import com.teamdev.jxbrowser.chromium.DefaultLoadHandler;
-import com.teamdev.jxbrowser.chromium.LoadParams;
+import com.teamdev.jxbrowser.chromium.BrowserPreferences;
 import com.teamdev.jxbrowser.chromium.swing.BrowserView;
 
 import javax.swing.*;
 import java.awt.*;
 
-/**
- * This example demonstrates how to cancel loading of a specific URL.
- */
-public class LoadHandlerExample {
+public class DisableSwipeNavigation {
     public static void main(String[] args) {
+        // Disables back/forward on left/right swipe on touch screens.
+        BrowserPreferences.setChromiumSwitches("--overscroll-history-navigation=0");
+
         Browser browser = new Browser();
         BrowserView view = new BrowserView(browser);
 
@@ -41,12 +40,6 @@ public class LoadHandlerExample {
         frame.setLocationRelativeTo(null);
         frame.setVisible(true);
 
-        browser.setLoadHandler(new DefaultLoadHandler() {
-            public boolean onLoad(LoadParams params) {
-                // Cancel loading URL that starts with http://www.google
-                return params.getURL().startsWith("http://www.google");
-            }
-        });
-        browser.loadURL("http://www.google.com");
+        browser.loadURL("http://google.com");
     }
 }
